search_service/requirements.txt
fastapi
uvicorn
qdrant-client
google-genai
python-dotenv
fastembed
